+++ Project_DXR/src/Game.cpp
@@ -473,7 +473,7 @@ void Game::update(double dt) {
 
 
 	// Camera movement
-	m_persCameraController->update(dt);
+	m_persCameraController->update(float(dt));
 
 	m_persCamera->updateConstantBuffer();
 	//m_meshes[0]->setTransform(t); // Updates transform matrix for rasterisation
@@ -500,7 +500,7 @@ void Game::update(double dt) {
 	if (m_animationSpeed > 0.0f) {
 		for (int i = 0; i < m_gameObjects.size(); i++) {
 			PotatoModel* pModel = m_gameObjects[i].getModel();
-			m_gameObjects[i].update(dt * m_animationSpeed);
+			m_gameObjects[i].update(float(dt) * m_animationSpeed);
 			m_meshes[m_animatedModelsStartIndex + i]->setTransform(m_gameObjects[i].getTransform());
 
 			if (m_dxRenderer->isDXRSupported())
@@ -540,8 +540,8 @@ void Game::render(double dt) {
 		return timerDt * timestampToMs;
 	};
 
-	m_timerSaver.addResult("BLAS", m_gameObjects.size(), getMsTime(Timers::BLAS));
-	m_timerSaver.addResult("TLAS", m_gameObjects.size(), getMsTime(Timers::TLAS));
+	m_timerSaver.addResult("BLAS", int(m_gameObjects.size()), getMsTime(Timers::BLAS));
+	m_timerSaver.addResult("TLAS", int(m_gameObjects.size()), getMsTime(Timers::TLAS));
 
 	//std::cout << "GPU time to update BLAS: " << timeInMs << std::endl;
 
@@ -584,7 +584,7 @@ void Game::imguiFunc() {
 					XMVECTOR vec;
 					vec = mesh->getTransform().getTranslationVec();
 					Transform& transform = mesh->getTransform();
-					if (i >= m_animatedModelsStartIndex) {
+					if (i >= UINT(m_animatedModelsStartIndex)) {
 						transform = m_gameObjects[i - m_animatedModelsStartIndex].getTransform();
 					}
 					if (ImGui::DragFloat3("Position", (float*)&vec, 0.1f)) {
